MCLOUD-5878: Create schema.error.yaml file with all errors and solutions

+++ src/Step/Deploy/RemoveDeployFailedFlag.php
@@ -9,9 +9,11 @@ namespace Magento\MagentoCloud\Step\Deploy;
 
 use Magento\MagentoCloud\Filesystem\Driver\File;
 use Magento\MagentoCloud\Filesystem\FileList;
+use Magento\MagentoCloud\Filesystem\FileSystemException;
 use Magento\MagentoCloud\Filesystem\Flag\Manager;
 use Magento\MagentoCloud\Step\StepException;
 use Magento\MagentoCloud\Step\StepInterface;
+use Magento\MagentoCloud\App\Logger\Error\ReaderInterface;
 
 /**
  * Removes flags and files which could be set during the previous deploy phase.
@@ -34,18 +36,26 @@ class RemoveDeployFailedFlag implements StepInterface
     private $fileList;
 
     /**
+     * @var ReaderInterface
+     */
+    private $errorReader;
+
+    /**
      * @param Manager $manager
      * @param File $fileDriver
      * @param FileList $fileList
+     * @param ReaderInterface $errorReader
      */
     public function __construct(
         Manager $manager,
         File $fileDriver,
-        FileList $fileList
+        FileList $fileList,
+        ReaderInterface $errorReader
     ) {
         $this->manager = $manager;
         $this->fileDriver = $fileDriver;
         $this->fileList = $fileList;
+        $this->errorReader = $errorReader;
     }
 
     /**
@@ -57,9 +67,31 @@ class RemoveDeployFailedFlag implements StepInterface
             $this->manager->delete(Manager::FLAG_DEPLOY_HOOK_IS_FAILED);
             $this->manager->delete(Manager::FLAG_IGNORE_SPLIT_DB);
             $this->manager->delete(Manager::FLAG_ENV_FILE_ABSENCE);
-            $this->fileDriver->deleteFile($this->fileList->getCloudErrorLog());
+
+            if ($this->isNeedToRemoveCloudErrorLog()) {
+                $this->fileDriver->deleteFile($this->fileList->getCloudErrorLog());
+            }
         } catch (\Exception $e) {
             throw new StepException($e->getMessage(), $e->getCode(), $e);
         }
+    }
+
+    /**
+     * Checks if cloud.error.log file should be removed.
+     * Returns true if file contains not only errors from the build phase
+     *
+     * @return bool
+     * @throws FileSystemException
+     */
+    private function isNeedToRemoveCloudErrorLog(): bool
+    {
+        $errors = $this->errorReader->read();
+        foreach ($errors as $error) {
+            if (isset($error['stage']) && $error['stage'] != 'build') {
+                return true;
+            }
+        }
+
+        return false;
     }
 }
